+++ feline/Element.h
@@ -14,19 +14,19 @@ class Element
 	Matrix3d undeformShapeMat,
 		     undeformShapeMatInv;
 
-	GenMatrix<double,6,12> strainMat;
-	GenMatrix<double,6,6> matConstantsMat;
-	GenMatrix<double,12,12> undeformStiffnessMat,
+	GenMatrix<float,6,12> strainMat;
+	GenMatrix<float,6,6> matConstantsMat;
+	GenMatrix<float,12,12> undeformStiffnessMat,
 						   massMat, RK, RKRT;
-	double undeformVolume;
-	double E, v; // E is Young's Modulus, v is Poisson's Ratio
-	double density;
+	float undeformVolume;
+	float E, v; // E is Young's Modulus, v is Poisson's Ratio
+	float density;
 	void preCompute();
 	void preComputeUndeformedStiffnessMat();
 	void preComputeMassMat();
 
 public:
-	Element(Node* n1, Node* n2, Node* n3, Node* n4, double _E, double _v, double _density);
+	Element(Node* n1, Node* n2, Node* n3, Node* n4, float _E, float _v, float _density);
 	Matrix3d computeDeformationMat();
 	Matrix3d computeDeformShapeMat();
 
@@ -36,9 +36,9 @@ public:
 	//computeundeformstiffnessmat B^T * c * B where B = LN
 	//computestiffnessmat dfe/dx |x = x*
 
-	GenMatrix<double,12,12>* getStiffnessMat(){return &undeformStiffnessMat;}
-	GenMatrix<double,12,12>* getMassMat(){return &massMat;}
-	void getRKRTandRK(GenMatrix<double,12,12>& RK, GenMatrix<double,12,12>& RKRT);
+	GenMatrix<float,12,12>* getStiffnessMat(){return &undeformStiffnessMat;}
+	GenMatrix<float,12,12>* getMassMat(){return &massMat;}
+	void getRKRTandRK(GenMatrix<float,12,12>& RK, GenMatrix<float,12,12>& RKRT);
 
 	void renderElement();
