+++ feline/Mesh.h
@@ -10,16 +10,16 @@ public:
 	QuickArray<Element*,MAX_ELEMENTS> elements;
 	int nodeIndices[MAX_ELEMENTS][4];
 	QuickArray<Node*,MAX_NODES> nodes;
-	double **globalStiffness, **globalMass;
+	float **globalStiffness, **globalMass;
 
 	int getNoOfElements(){return elements.size();}
 	int getNoOfNodes(){return numnodes;}
 	
-	void addElement(int ind0, int ind1, int ind2, int ind3, double _E, double _v, double density);
+	void addElement(int ind0, int ind1, int ind2, int ind3, float _E, float _v, float density);
 	void resetGlobalStiffness();
 	void resetGlobalMass();
-	double** assembleGlobalMass();
-	double** assembleGlobalStiffness();
+	float** assembleGlobalMass();
+	float** assembleGlobalStiffness();
 	Mesh(Node nodes[], int n);
 	~Mesh(void);
 };
