+++ feline/Node.h
@@ -6,9 +6,9 @@ class Node
 {
 	//implements a node
 public:
-	double mass;
-	vector3<double> pos, pos_t, vec_t,force;
-	Node(vector3<double> _pos, vector3<double> _vel, vector3<double> _force, double _mass);
+	float mass;
+	vector3<float> pos, pos_t, vec_t,force;
+	Node(vector3<float> _pos, vector3<float> _vel, vector3<float> _force, float _mass);
 	Node(void);
 	~Node(void);
 };
