+++ feline/PolarDecompose.cpp
@@ -1,24 +1,24 @@
 #include "PolarDecompose.h"
 
 
-double PolarDecompose::oneNorm(Matrix3d& F)
+float PolarDecompose::oneNorm(Matrix3d& F)
 {
-  double norm = 0.0;
+  float norm = 0.0;
   for (int i=0; i<3; i++) 
   {
-    double columnAbsSum = fabs(F(0,i)) + fabs(F(1,i)) + fabs(F(2,i));
+    float columnAbsSum = fabs(F(0,i)) + fabs(F(1,i)) + fabs(F(2,i));
     if (columnAbsSum > norm) 
       norm = columnAbsSum;
   }
   return norm;
 }
 
-double PolarDecompose::infNorm(Matrix3d& F)
+float PolarDecompose::infNorm(Matrix3d& F)
 {
-  double norm = 0.0;
+  float norm = 0.0;
   for (int i=0; i<3; i++) 
   {
-    double rowSum = fabs(F(i,0)) + fabs(F(i,1)) + fabs(F(i,2));
+    float rowSum = fabs(F(i,0)) + fabs(F(i,1)) + fabs(F(i,2));
     if (rowSum > norm) 
       norm = rowSum;
   }
@@ -29,7 +29,7 @@ void PolarDecompose::compute(Matrix3d& F, Matrix3d& R, Matrix3d& S)
 {
   Matrix3d Mk;
   Matrix3d Ek;
-  double det, M_oneNorm, M_infNorm, E_oneNorm;
+  float det, M_oneNorm, M_infNorm, E_oneNorm;
 
   Mk = F.transpose();
 
@@ -39,9 +39,9 @@ void PolarDecompose::compute(Matrix3d& F, Matrix3d& R, Matrix3d& S)
   do 
   {
    
-	vector3<double> r1 = vector3<double>(Mk(1,0), Mk(1,1), Mk(1,2)).cross(vector3<double>(Mk(2,0), Mk(2,1), Mk(2,2))); 
-	vector3<double> r2 = vector3<double>(Mk(2,0), Mk(2,1), Mk(2,2)).cross(vector3<double>(Mk(0,0), Mk(0,1), Mk(0,2))); 
-	vector3<double> r3 = vector3<double>(Mk(0,0), Mk(0,1), Mk(0,2)).cross(vector3<double>(Mk(1,0), Mk(1,1), Mk(1,2))); 
+	vector3<float> r1 = vector3<float>(Mk(1,0), Mk(1,1), Mk(1,2)).cross(vector3<float>(Mk(2,0), Mk(2,1), Mk(2,2))); 
+	vector3<float> r2 = vector3<float>(Mk(2,0), Mk(2,1), Mk(2,2)).cross(vector3<float>(Mk(0,0), Mk(0,1), Mk(0,2))); 
+	vector3<float> r3 = vector3<float>(Mk(0,0), Mk(0,1), Mk(0,2)).cross(vector3<float>(Mk(1,0), Mk(1,1), Mk(1,2))); 
    Matrix3d MadjTk(r1.x,r1.y,r1.z,
 					r2.x,r2.y,r2.z,
 					r3.x,r3.y,r3.z);
@@ -54,12 +54,12 @@ void PolarDecompose::compute(Matrix3d& F, Matrix3d& R, Matrix3d& S)
       break;
     }
 
-    double MadjT_one = oneNorm(MadjTk); 
-    double MadjT_inf = infNorm(MadjTk);
+    float MadjT_one = oneNorm(MadjTk); 
+    float MadjT_inf = infNorm(MadjTk);
 
-    double gamma = sqrt(sqrt((MadjT_one * MadjT_inf) / (M_oneNorm * M_infNorm)) / fabs(det));
-    double g1 = gamma * 0.5;
-    double g2 = 0.5 / (gamma * det);
+    float gamma = sqrt(sqrt((MadjT_one * MadjT_inf) / (M_oneNorm * M_infNorm)) / fabs(det));
+    float g1 = gamma * 0.5;
+    float g2 = 0.5 / (gamma * det);
 
    for(int i=0; i<3; i++)
     for(int j=0; j<3; j++)
