+++ feline/PolarDecompose.h
@@ -6,8 +6,8 @@
 class PolarDecompose
 {
 public:
-	static double oneNorm(Matrix3d& F);
-	static double infNorm(Matrix3d& F);
+	static float oneNorm(Matrix3d& F);
+	static float infNorm(Matrix3d& F);
 	static void compute(Matrix3d& F, Matrix3d& R, Matrix3d& S);
 };
